Move profunctor code to separate module.
