+++ optics-core/optics-core.cabal
@@ -15,3 +15,4 @@ library
     Optics
   other-modules:
     Optics.Internal
+    Optics.Internal.Profunctor
